+++ src/cones/coneops_socone.jl
@@ -6,13 +6,17 @@
 degree(K::SecondOrderCone{T}) where {T} = 1
 numel(K::SecondOrderCone{T}) where {T} = K.dim
 
+function is_sparse_expandable(K::SecondOrderCone{T}) where{T}
+    return !isnothing(K.sparse_data)
+end
+
 function margins(
     K::SecondOrderCone{T},
     z::AbstractVector{T},
     pd::PrimalOrDualCone
 ) where{T}
 
-    α = z[1] - norm(z[2:end])
+    @views α = z[1] - norm(z[2:end])
     β = max(zero(T),α)
     (α,β)
 end
@@ -54,11 +58,16 @@ function set_identity_scaling!(
     K::SecondOrderCone{T}
 ) where {T}
 
-    K.d  = one(T)
-    K.u .= zero(T)
-    K.v .= zero(T)
-    K.η  = one(T)
     K.w .= zero(T)
+    K.w[1] = one(T)
+    K.η  = one(T)
+
+    if !isnothing(K.sparse_data)
+        K.sparse_data.d  = T(0.5)
+        K.sparse_data.u .= zero(T)
+        K.sparse_data.u[1] = sqrt(T(0.5))
+        K.sparse_data.v .= zero(T)
+    end 
 
     return nothing
 end
@@ -72,21 +81,27 @@ function update_scaling!(
 ) where {T}
 
     #first calculate the scaled vector w
-    @views zscale = _sqrt_soc_residual(z)
-    @views sscale = _sqrt_soc_residual(s)
+    zscale = _sqrt_soc_residual(z)
+    sscale = _sqrt_soc_residual(s)
 
     # Fail if either s or z is not an interior point
     if iszero(zscale) || iszero(sscale)
         return is_scaling_success = false
     end
 
+    #the leading scalar term for W^TW
+    K.η = sqrt(sscale/zscale)
+
     # construct w and normalize
     w = K.w
     w     .= s./(sscale)
     w[1]  += z[1]/(zscale)
-    @views w[2:end] .-= z[2:end]/(zscale)
-    wscale = _sqrt_soc_residual(w)
 
+    @inbounds for i in 2:length(w)
+        w[i] -= z[i]/(zscale)
+    end
+
+    wscale = _sqrt_soc_residual(w)
     # Fail if w is not an interior point
     if iszero(wscale)
         return is_scaling_success = false
@@ -94,35 +109,46 @@ function update_scaling!(
     w ./= wscale
 
     #try to force badly scaled w to come out normalized
-    @views w1sq = dot(w[2:end],w[2:end])
+    @views w1sq = sumsq(w[2:end])
     w[1] = sqrt(1 + w1sq)
-    wsq    = w[1]*w[1] + w1sq
 
-    #various intermediate calcs for u,v,d,η
-    α  = 2*w[1]
+    #Compute the scaling point λ.   Should satisfy λ = Wz = W^{-T}s
+    γ = 0.5 * wscale
+    K.λ[1] = γ 
+    s1 = @view s[2:end];
+    z1 = @view z[2:end];
+    λ1 = @view K.λ[2:end];
+    λ1 .= ((γ + z[1]/zscale)/sscale).*s1 +((γ + s[1]/sscale)/zscale).*z1 
+    K.λ[2:end] *= inv(s[1]/sscale + z[1]/zscale + 2*γ)
+    K.λ .*= sqrt(sscale*zscale)
 
-    #Scalar d is the upper LH corner of the diagonal
-    #term in the rank-2 update form of W^TW
-    wsqinv = 1/wsq
-    K.d    = wsqinv / 2
+    #Populate sparse expansion terms if allocated
+    if is_sparse_expandable(K)
 
-    #the leading scalar term for W^TW
-    K.η = sqrt(sscale/zscale)
+        sparse_data = K.sparse_data
 
-    #the vectors for the rank two update
-    #representation of W^TW
-    u0  = sqrt(wsq - K.d)
-    u1 = α/u0
-    v0 = zero(T)
-    v1 = sqrt( 2*(2 + wsqinv)/(2*wsq - wsqinv))
-    
-    K.u[1] = u0
-    @views K.u[2:end] .= u1.*K.w[2:end]
-    K.v[1] = v0
-    @views K.v[2:end] .= v1.*K.w[2:end]
+        #various intermediate calcs for u,v,d
+        α  = 2*w[1]
 
-    #λ = Wz
-    mul_W!(K,:N,K.λ,z,one(T),zero(T))
+        #Scalar d is the upper LH corner of the diagonal
+        #term in the rank-2 update form of W^TW
+        wsq    = w[1]*w[1] + w1sq
+        wsqinv = 1/wsq
+        sparse_data.d    = wsqinv / 2
+
+        #the vectors for the rank two update
+        #representation of W^TW
+        u0  = sqrt(wsq - sparse_data.d)
+        u1 = α/u0
+        v0 = zero(T)
+        v1 = sqrt( 2*(2 + wsqinv)/(2*wsq - wsqinv))
+        
+        sparse_data.u[1] = u0
+        @views K.sparse_data.u[2:end] .= u1.*w[2:end]
+        sparse_data.v[1] = v0
+        @views sparse_data.v[2:end] .= v1.*w[2:end]
+
+    end 
 
     return is_scaling_success = true
 end
@@ -132,16 +158,43 @@ function get_Hs!(
     Hsblock::AbstractVector{T}
 ) where {T}
 
-    #NB: we are returning here the diagonal D block from the
-    #sparse representation of W^TW, but not the
-    #extra two entries at the bottom right of the block.
-    #The ConicVector for s and z (and its views) don't
-    #know anything about the 2 extra sparsifying entries
+    if is_sparse_expandable(K)
+        #For sparse form, we are returning here the diagonal D block 
+        #from the sparse representation of W^TW, but not the
+        #extra two entries at the bottom right of the block.
+        #The ConicVector for s and z (and its views) don't
+        #know anything about the 2 extra sparsifying entries
+        Hsblock    .= K.η^2
+        Hsblock[1] *= K.sparse_data.d
+    
+    else 
+        #for dense form, we return H = \eta^2 (2*ww^T - J), where 
+        #J = diag(1,-I).  We are packing into dense triu form
 
-    Hsblock    .= (K.η^2)
-    Hsblock[1] *= K.d
+        Hsblock[1] = 2*K.w[1]^2 - one(T)
+        hidx = 2
+
+        @inbounds for col in 2:K.dim
+            wcol = K.w[col]
+            @inbounds for row in 1:col 
+                Hsblock[hidx] = 2*K.w[row]*wcol
+                hidx += 1
+            end 
+            #go back to add the offset term from J 
+            Hsblock[hidx-1] += one(T)
+        end
+        Hsblock .*= K.η^2
+    end
 
     return nothing
+
+    
+end
+
+function Hs_is_diagonal(
+    K::SecondOrderCone{T}
+) where{T}
+    return is_sparse_expandable(K)
 end
 
 # compute the product y = WᵀWx
@@ -152,9 +205,13 @@ function mul_Hs!(
     work::AbstractVector{T}
 ) where {T}
 
-    mul_W!(K,:N,work,x,one(T),zero(T))    #work = Wx
-    mul_W!(K,:T,y,work,one(T),zero(T))    #y = c Wᵀwork = W^TWx
-  
+    # y = = H^{-1}x = W^TWx
+    # where H^{-1} = \eta^{-2} (2*ww^T - J)
+    c = 2*dot(K.w,x)
+    y .= x
+    y[1] = -x[1]
+    y .+= c*K.w
+    y .*= K.η^2  
     return nothing
 end
 
@@ -188,9 +245,26 @@ function Δs_from_Δz_offset!(
     work::AbstractVector{T},
     z::AbstractVector{T}
 ) where {T}
- 
-    #Wᵀ(λ \ ds)
-       _Δs_from_Δz_offset_symmetric!(K,out,ds,work)
+   
+    #out = Wᵀ(λ \ ds).  Below is equivalent,
+    #but appears to be a little more stable 
+    
+    resz = _soc_residual(z)
+
+    @views λ1ds1  = dot(K.λ[2:end],ds[2:end])
+    @views w1ds1  = dot(K.w[2:end],ds[2:end])
+
+    out .= -z
+    out[1] = +z[1]
+
+    c = K.λ[1]*ds[1] - λ1ds1
+    out .*= c/resz
+
+    out[1]              += K.η*w1ds1
+    @views out[2:end]  .+= K.η*(ds[2:end] + w1ds1/(1+K.w[1]).*K.w[2:end])
+
+    out .*= (1/K.λ[1])
+
 end
 
 #return maximum allowable step length while remaining in the socone
@@ -252,7 +326,10 @@ function mul_W!(
   c = x[1] + ζ/(1+K.w[1])
 
   y[1] = α*K.η*(K.w[1]*x[1] + ζ) + β*y[1]
-  @views y[2:end] .= (α*K.η)*(x[2:end] + c*K.w[2:end]) + β*y[2:end]
+
+  @inbounds for i in 2:length(y)
+      y[i] = α*K.η*(x[i] + c*K.w[i]) + β*y[i]
+  end
 
   return nothing
 end
@@ -275,7 +352,7 @@ function mul_Winv!(
 
     y[1] = (α/K.η)*(K.w[1]*x[1] - ζ) + β*y[1]
 
-    for i = 2:length(y)
+    @inbounds for i = 2:length(y)
         y[i] = (α/K.η)*(x[i] + c*K.w[i]) + β*y[i]
     end
 
@@ -329,7 +406,7 @@ function inv_circ_op!(
     @views v = dot(y[2:end],z[2:end])
 
     x[1]      = (y[1]*z[1] - v)*pinv
-    @views x[2:end] .= pinv*(v/y[1] - z[1]).*y[2:end] + (1/y[1]).*z[2:end]
+    @views x[2:end] .= pinv.*(v/y[1] - z[1]).*y[2:end] + (1/y[1]).*z[2:end]
 
     return nothing
 end
@@ -340,7 +417,7 @@ end
 
 @inline function _soc_residual(z:: AbstractVector{T}) where {T} 
 
-    @views z[1]*z[1] - dot(z[2:end],z[2:end])
+    @views z[1]*z[1] - sumsq(z[2:end])
 end 
 
 @inline function _sqrt_soc_residual(z:: AbstractVector{T}) where {T}