+++ src/kktsolvers/direct-ldl/includes.jl
@@ -2,7 +2,8 @@
 include("./directldl_defaults.jl")
 include("./directldl_qdldl.jl")
 include("./directldl_cholmod.jl")
-include("./directldl_utils.jl")
+include("./directldl_kkt_assembly.jl")
+include("./directldl_datamaps.jl")
 
 #NB: MKL is an optional dependency and is 
 #loaded via Requires.jl in the main Clarabel module