+++ src/kktsolvers/kktsolver_directldl.jl
@@ -45,9 +45,15 @@ mutable struct DirectLDLKKTSolver{T} <: AbstractKKTSolver{T}
 
     function DirectLDLKKTSolver{T}(P,A,cones,m,n,settings) where {T}
 
-        #solving in sparse format.  Need this many
-        #extra variables for SOCs
-        p = 2*cones.type_counts[SecondOrderCone]
+        # get a constructor for the LDL solver we should use,
+        # and also the matrix shape it requires
+        (kktshape, ldlsolverT) = _get_ldlsolver_config(settings)
+
+        #construct a KKT matrix of the right shape
+        KKT, map = _assemble_kkt_matrix(P,A,cones,kktshape)
+
+        #Need this many extra variables for sparse cones
+        p = pdim(map.sparse_maps)
 
         #LHS/RHS/work for iterative refinement
         x    = Vector{T}(undef,n+m+p)
@@ -56,19 +62,12 @@ mutable struct DirectLDLKKTSolver{T} <: AbstractKKTSolver{T}
         work_dx = Vector{T}(undef,n+m+p)
 
         #the expected signs of D in LDL
-        Dsigns = Vector{Int}(undef,n+m+p)
-        _fill_Dsigns!(Dsigns,m,n,p)
+        Dsigns = Vector{Int}(undef,n+m+p)     
+        _fill_Dsigns!(Dsigns,m,n,map)
 
         #updates to the diagonal of KKT will be
         #assigned here before updating matrix entries
         Hsblocks = _allocate_kkt_Hsblocks(T, cones)
-
-        #which LDL solver should I use?
-        ldlsolverT = _get_ldlsolver_type(settings.direct_solve_method)
-
-        #does it want a :triu or :tril KKT matrix?
-        kktshape = required_matrix_shape(ldlsolverT)
-        KKT, map = _assemble_kkt_matrix(P,A,cones,kktshape)
 
         diagonal_regularizer = zero(T)
 
@@ -98,16 +97,32 @@ function _get_ldlsolver_type(s::Symbol)
     end
 end
 
-function _fill_Dsigns!(Dsigns,m,n,p)
+function _get_ldlsolver_config(settings::Settings)
+
+    #which LDL solver should I use?
+    ldlsolverT = _get_ldlsolver_type(settings.direct_solve_method)
+
+    #does it want a :triu or :tril KKT matrix?
+    kktshape = required_matrix_shape(ldlsolverT)
+
+    (kktshape,ldlsolverT)
+end 
+
+
+function _fill_Dsigns!(Dsigns,m,n,map)
 
     Dsigns .= 1
 
     #flip expected negative signs of D in LDL
     Dsigns[n+1:n+m] .= -1
 
-    #the trailing block of p entries should
-    #have alternating signs
-    Dsigns[(n+m+1):2:(n+m+p)] .= -1
+    p = m + n + 1
+    # assign D signs for sparse expansion cones
+    for thismap in map.sparse_maps
+        thisp = pdim(thismap)
+        Dsigns[p:(p+thisp-1)] .= Clarabel.Dsigns(thismap)   
+        p += thisp
+    end
 end
 
 #update entries in the kktsolver object using the
@@ -115,8 +130,8 @@ end
 function _update_values!(
     ldlsolver::AbstractDirectLDLSolver{T},
     KKT::SparseMatrixCSC{T,Ti},
-    index::Vector{Ti},
-    values::Vector{T}
+    index::AbstractVector{Ti},
+    values::AbstractVector{T}
 ) where{T,Ti}
 
     #Update values in the KKT matrix K
@@ -132,8 +147,8 @@ end
 #updates KKT matrix values
 function _update_values_KKT!(
     KKT::SparseMatrixCSC{T,Int},
-    index::Vector{Ti},
-    values::Vector{T}
+    index::AbstractVector{Ti},
+    values::AbstractVector{T}
 ) where{T,Ti}
 
     #Update values in the KKT matrix K
@@ -146,7 +161,7 @@ end
 function _scale_values!(
     ldlsolver::AbstractDirectLDLSolver{T},
     KKT::SparseMatrixCSC{T,Ti},
-    index::Vector{Ti},
+    index::AbstractVector{Ti},
     scale::T
 ) where{T,Ti}
 
@@ -163,7 +178,7 @@ end
 #updates KKT matrix values
 function _scale_values_KKT!(
     KKT::SparseMatrixCSC{T,Int},
-    index::Vector{Ti},
+    index::AbstractVector{Ti},
     scale::T
 ) where{T,Ti}
 
@@ -195,8 +210,6 @@ function _kktsolver_update_inner!(
 
     #real implementation is here, and now ldlsolver
     #will be compiled to something concrete.
-
-    settings  = kktsolver.settings
     map       = kktsolver.map
     KKT       = kktsolver.KKT
 
@@ -205,31 +218,21 @@ function _kktsolver_update_inner!(
 
     for (index, values) in zip(map.Hsblocks,kktsolver.Hsblocks)
         #change signs to get -W^TW
-        # values .= -values
         @. values *= -one(T)
         _update_values!(ldlsolver,KKT,index,values)
     end
 
-    #update the scaled u and v columns.
-    cidx = 1        #which of the SOCs are we working on?
+    sparse_map_iter = Iterators.Stateful(map.sparse_maps)
 
-    for (i,cone) = enumerate(cones)
-        if isa(cone,SecondOrderCone)
-            η2 = cone.η^2
+    updateFcn = (index,values) -> _update_values!(ldlsolver,KKT,index,values)
+    scaleFcn  = (index,scale)  -> _scale_values!(ldlsolver,KKT,index,scale)
 
-            #off diagonal columns (or rows)
-            _update_values!(ldlsolver,KKT,map.SOC_u[cidx],cone.u)
-            _update_values!(ldlsolver,KKT,map.SOC_v[cidx],cone.v)
-            _scale_values!(ldlsolver,KKT,map.SOC_u[cidx],-η2)
-            _scale_values!(ldlsolver,KKT,map.SOC_v[cidx],-η2)
-
-
-            #add η^2*(1/-1) to diagonal in the extended rows/cols
-            _update_values!(ldlsolver,KKT,[map.SOC_D[cidx*2-1]],[-η2])
-            _update_values!(ldlsolver,KKT,[map.SOC_D[cidx*2  ]],[+η2])
-
-            cidx += 1
-        end
+    for cone in cones
+        #add sparse expansions columns for sparse cones 
+        if @conedispatch is_sparse_expandable(cone)
+            thismap = popfirst!(sparse_map_iter)
+            _csc_update_sparsecone(cone,thismap,updateFcn,scaleFcn)
+        end 
     end
 
     return _kktsolver_regularize_and_refactor!(kktsolver, ldlsolver)
@@ -326,7 +329,7 @@ function kktsolver_getlhs!(
 ) where {T}
 
     x = kktsolver.x
-    (m,n,p) = (kktsolver.m,kktsolver.n,kktsolver.p)
+    (m,n) = (kktsolver.m,kktsolver.n)
 
     isnothing(lhsx) || (@views lhsx .= x[1:n])
     isnothing(lhsz) || (@views lhsz .= x[(n+1):(n+m)])